backend/requirements.txt
# FastAPI Framework
fastapi==0.115.5
uvicorn[standard]==0.32.1

# Supabase Client
supabase==2.10.0

# Environment Variables
python-dotenv==1.0.1

# Data Validation
pydantic==2.10.3
pydantic-settings==2.6.1

# CORS
fastapi-cors==0.0.6

# HTTP Client (compatible con supabase 2.10.0)
httpx>=0.26,<0.28
